rename cirruslib to cirrus.lib
